+++ libs/cosmos-sdk/x/auth/types/stdtx.go
@@ -69,7 +69,7 @@ func (tx *StdTx) UnmarshalFromAmino(cdc *amino.Codec, data []byte) error {
 			return err
 		}
 		data = data[n:]
-		if len(data) < int(dataLen) {
+		if uint64(len(data)) < dataLen {
 			return fmt.Errorf("invalid tx data")
 		}
 		subData = data[:dataLen]
@@ -364,7 +364,7 @@ func (fee *StdFee) UnmarshalFromAmino(cdc *amino.Codec, data []byte) error {
 				return err
 			}
 			data = data[n:]
-			if len(data) < int(dataLen) {
+			if uint64(len(data)) < dataLen {
 				return fmt.Errorf("invalid tx data")
 			}
 			subData = data[:dataLen]
@@ -535,7 +535,7 @@ func (ss *StdSignature) UnmarshalFromAmino(cdc *amino.Codec, data []byte) error 
 			return err
 		}
 		data = data[n:]
-		if len(data) < int(dataLen) {
+		if uint64(len(data)) < dataLen {
 			return fmt.Errorf("invalid tx data")
 		}
 		subData = data[:dataLen]
